requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
discord-webhook==1.4.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
Telethon==1.39.0
urllib3==2.3.0
